fix "max depth exceeded" error when building image
The maximum number of layers of the docker image is 128.  The final
product will reach this limit because it will sequentially copy the
python binaries from many images.

To solve this problem, I changed to an approach that built all versions
in single image.  The disadvantage of this approach is that it takes lot
of time to troubleshooting of build failure.